organize ricorrenti types
