--- Frontend-nextjs/app/(protected)/transazioni/components/list/table/types.ts
@@ -1,5 +1,0 @@
-// types.ts
-import { Transaction } from "@/types/types/transaction";
-
-// Estende Transaction con campo di raggruppamento
-export type TransactionWithGroup = Transaction & { monthGroup: string };
