+++ Frontend-nextjs/types/ricorrenti/card/index.ts
@@ -1,0 +1,10 @@
+import type { Ricorrenza } from "@/types/types/ricorrenza";
+
+export type CardTotaliAnnuiProps = {
+    ricorrenze: Ricorrenza[];
+};
+
+export type CardGraficoPagamentiProps = {
+    ricorrenze: Ricorrenza[];
+    customDaysArr?: Date[];
+};
