+++ Frontend-nextjs/types/ricorrenti/index.ts
@@ -1,0 +1,3 @@
+// Tipi specifici per la pagina Ricorrenti
+export * from "./card";
+export * from "./liste";
