+++ Frontend-nextjs/types/ricorrenti/liste/components/index.ts
@@ -1,0 +1,15 @@
+import type { Ricorrenza } from "@/types/types/ricorrenza";
+
+export type RicorrenzaGroupProps = {
+    freq: string;
+    items: Ricorrenza[];
+    showSeparator?: boolean;
+    onEdit?: (r: Ricorrenza) => void;
+    onDelete?: (r: Ricorrenza) => void;
+};
+
+export type RicorrenzaItemProps = {
+    r: Ricorrenza;
+    onEdit?: (r: Ricorrenza) => void;
+    onDelete?: (r: Ricorrenza) => void;
+};
