+++ Frontend-nextjs/types/transazioni/index.ts
@@ -1,0 +1,3 @@
+// Tipi specifici per la pagina Transazioni
+export * from "./list";
+export * from "./modal";
